requirements.txt
asgiref==3.5.2
azure-core==1.26.1
azure-storage-blob==12.14.1
backports.zoneinfo==0.2.1
boto3==1.26.3
botocore==1.29.3
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
decorator==5.1.1
Django==4.1.3
django-cors-headers==3.13.0
django-extensions==3.2.1
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-storages==1.13.1
djangorestframework==3.14.0
dynaconf==3.1.11
hvac==1.0.2
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
jmespath==1.0.1
knox==0.1.14
loguru==0.6.0
MarkupSafe==2.1.1
msrest==0.7.1
oauthlib==3.2.2
Pillow==9.3.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
pyhcl==0.4.4
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.12
validators==0.20.0
whitenoise==6.2.0
